add isAuthetcated check for several use cases with tests for them

+++ src/main/kotlin/logic/usecase/task/DeleteTaskUseCase.kt
@@ -1,10 +1,12 @@
 package org.baghdad.logic.usecase.task
 
+import org.baghdad.logic.manager.SessionManager
 import org.baghdad.logic.model.entities.Action
 import org.baghdad.logic.model.entities.AuditLogEntity
 import org.baghdad.logic.model.enums.Entities
 import org.baghdad.logic.model.entities.TaskEntity
 import org.baghdad.logic.model.entities.UserEntity
+import org.baghdad.logic.model.exceptions.UnauthorizedException
 import org.baghdad.logic.repositories.AuditRepository
 import org.baghdad.logic.repositories.TaskRepository
 import org.baghdad.logic.repositories.UserRepository
@@ -13,10 +15,12 @@ import java.util.UUID
 class DeleteTaskUseCase(
     private val taskRepository: TaskRepository,
     private val auditRepository: AuditRepository,
-    private val userRepository: UserRepository
+    private val userRepository: UserRepository,
+    private val sessionManager: SessionManager
 ) {
 
     suspend operator fun invoke(taskId: UUID, userId: UUID) {
+        if (!sessionManager.isAuthenticated()) throw UnauthorizedException("Not authenticated")
         val task = taskRepository.getTaskById(taskId)
         taskRepository.deleteTask(taskId)
